+++ sdk/storage/storage-blob/src/index.browser.ts
@@ -4,9 +4,7 @@
 import { RestError } from "@azure/core-http";
 
 export * from "./BlobServiceClient";
-export * from "./ContainerClient";
-// ordering of these clients is important as there are circular dependencies
-export * from "./BlobClient";
+export * from "./Clients";
 export * from "./BlobBatch";
 export * from "./BlobBatchClient";
 export * from "./BatchResponse";
@@ -15,7 +13,6 @@ export * from "./credentials/AnonymousCredential";
 export * from "./credentials/Credential";
 export { SasIPRange } from "./SasIPRange";
 export { Range } from "./Range";
-export * from "./BlobLeaseClient";
 export { BlockBlobTier, PremiumPageBlobTier } from "./models";
 export * from "./Pipeline";
 export * from "./policies/AnonymousCredentialPolicy";
